Further cleanup add coverage report options
