+++ .gitignore
@@ -7,6 +7,7 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
+# coverage
 node_modules
 dist
 dist-ssr
