+++ package.json
@@ -10,6 +10,7 @@
     "lint:formatting": "prettier . --check",
     "format": "prettier . --write",
     "test": "vitest",
+    "test:coverage": "vitest --coverage --ui --open",
     "preview": "vite preview"
   },
   "dependencies": {
@@ -42,6 +43,7 @@
     "@typescript-eslint/parser": "^6.0.0",
     "@vitejs/plugin-react": "^4.0.3",
     "@vitest/coverage-v8": "^0.34.1",
+    "@vitest/ui": "^0.34.1",
     "eslint": "^8.45.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
